# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateProductRequest:
    properties:
      currency:
        maxLength: 3
        minLength: 3
        type: string
      price:
        type: number
      product_description:
        type: string
      product_name:
        minLength: 3
        type: string
      stock:
        type: integer
      user_id:
        type: string
    required:
    - currency
    - price
    - product_name
    - stock
    - user_id
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: boolean
      message:
        type: string
    type: object
...
  handler.ProductResponse:
    properties:
      message:
        type: string
      product:
        $ref: '#/definitions/model.Product'
    type: object
...
  handler.UpdateProductRequest:
    properties:
      currency:
        maxLength: 3
        minLength: 3
        type: string
      price:
        type: number
      product_description:
        type: string
      product_name:
        minLength: 3
        type: string
      stock:
        type: integer
    required:
    - currency
    - price
    - product_name
    - stock
    type: object
...
  model.Product:
    properties:
      created_at:
        type: string
      currency:
        type: string
      price:
        type: number
      product_code:
        type: string
      product_description:
        type: string
      product_name:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: This is a sample server Instashop server.
  title: Instashop Swagger API
  version: "1.0"
paths:
...
  /api/v1/admin/product:
    post:
      description: Add a new product to the database
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.ProductResponse'
            - properties:
                ' message':
                  type: string
                product:
                  $ref: '#/definitions/model.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /api/v1/admin/product/{product_code}:
    delete:
      description: Delete a product by its product code
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product Code
        in: path
        name: product_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product has been successfully deleted
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
        "404":
          description: Product not found
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
        "500":
          description: Error in deleting product
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
    put:
      description: Update product details
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product Code
        in: path
        name: product_code
        required: true
        type: string
      - description: Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.ProductResponse'
            - properties:
                ' message':
                  type: string
                product:
                  $ref: '#/definitions/model.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Update an existing product
      tags:
      - Products
  /api/v1/product/{product_code}:
    get:
      description: Retrieve product details using the product code
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product Code
        in: path
        name: product_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product successfully retrieved
          schema:
            allOf:
            - $ref: '#/definitions/handler.ProductResponse'
            - properties:
                ' message':
                  type: string
                product:
                  $ref: '#/definitions/model.Product'
              type: object
        "404":
          description: No product found
          schema:
            allOf:
            - $ref: '#/definitions/handler.ProductResponse'
            - properties:
                ' message':
                  type: string
                product:
                  $ref: '#/definitions/model.Product'
              type: object
        "500":
          description: Failed to retrieve product
          schema:
            allOf:
            - $ref: '#/definitions/handler.ErrorResponse'
            - properties:
                ' message':
                  type: string
                error:
                  type: boolean
              type: object
      security:
      - BearerAuth: []
      summary: Get a product by its product code
      tags:
      - Products
...
swagger: "2.0"